Energy now integer not double. No idea why I even thought of making it a double in the first place.

+++ Assets/Scripts/Environment/Foodbit.cs
@@ -27,7 +27,7 @@ public class Foodbit : MonoBehaviour {
 	
 	private Ether eth;
 	private MeshRenderer mr;
-	private double energy;
+	private int energy;
 	
 	
 	void Start () {
@@ -39,7 +39,7 @@ public class Foodbit : MonoBehaviour {
 		co.isTrigger = true;
 	}
 	
-	public double getEnergy() {
+	public int  getEnergy() {
 		return energy;
 	}
